Add new subclass for Lidl tickets
New ticket formats may require a different set of options for the slicer
in order to properly identify ticket lines. These options are specific
to each ticket format but need to be known by the slicer.

This commit introduces a new class SlicerOptions that wraps the previous
default options of the slicer. By default the slicer uses a predefined
set of options, unless a specific ticket defines any custom property.